refactor(db-dialogs): db_dialogs.py(4012줄)를 책임별 5개 파일로 분해
대형 파일 유지보수 리스크 해소 (TF-STATUS-006 watch 항목). 결합도가 거의 0에
가까운 4개 독립 단위(DB 연결 / Export / Import / 고아 레코드 분석)를 분리하고,
원본 db_dialogs.py는 RustDumpWizard + 전체 재수입 파사드로 축소.

- src/ui/dialogs/db_connection_dialog.py (신규): DBConnectionDialog
- src/ui/dialogs/db_export_dialog.py (신규): RustDumpExportDialog + export 헬퍼
- src/ui/dialogs/db_import_dialog.py (신규): RustDumpImportDialog + import/sanitize 헬퍼
- src/ui/dialogs/db_orphan_dialog.py (신규): OrphanRecordDialog + 2 워커 + 헬퍼
- src/ui/dialogs/db_dialogs.py: 4012 -> 216줄 (RustDumpWizard만 남고 나머지는 재수입)
- scripts/rust-core-regression-gate.ps1: MySQLConnector 직접 import하는 신규
  4개 파일을 allowlist에 추가

외부 소비자(main_window.py, migration_dialogs.py, dialogs/__init__.py)는 전부
db_dialogs.py를 통해 그대로 import 가능 — 파사드 재수입으로 코드 변경 불필요.

테스트도 동일 경계로 분리(test_db_export_dialog.py/test_db_import_dialog.py/
test_db_orphan_dialog.py 신규, test_db_dialogs.py는 RustDumpWizard 글루 코드
테스트만 남아 1638 -> 183줄). monkeypatch 문자열 경로를 실제 코드가 위치한
새 모듈로 갱신(예: db_dialogs.check_rust_dump -> db_export_dialog.check_rust_dump).

검증: 전체 pytest 1809 passed / 1 known-flaky(macOS CI 의존, 로컬 상시 실패)
제외 전부 통과. py_compile 전체 통과. 앱 스모크(core_hello) 성공.
