+++ scripts/rust-core-regression-gate.ps1
@@ -19,6 +19,10 @@ $allowedEngineLocked = @(
     "src/core/migration_preflight.py",
     "src/exporters/rust_dump_exporter.py",
     "src/ui/dialogs/db_dialogs.py",
+    "src/ui/dialogs/db_connection_dialog.py",
+    "src/ui/dialogs/db_export_dialog.py",
+    "src/ui/dialogs/db_import_dialog.py",
+    "src/ui/dialogs/db_orphan_dialog.py",
     "src/ui/dialogs/diff_dialog.py",
     "src/ui/dialogs/fix_wizard_dialog.py",
     "src/ui/dialogs/migration_dialogs.py",
